# improve search performance by scanning with the timestamp index iteratively